loan repayment program in C++
When you borrow money to buy a house, a car, or for some other purpose, you repay the loan by making periodic payments over a certain period of time. The lending company will charge interest on the loan. Every periodic payment consists of the interest on the loan and the payment toward the principal amount. This program calculates how many months it takes to repay a loan, validates user input, and adjusts the last payment if necessary.